Add files via upload
Farming has been the backbone of human civilization since the beginning of time. Farmers are responsible for feeding the world's population, and their efforts are critical in ensuring food security. However, selecting the best crop for a farm can be a daunting task, and farmers face several challenges in this regard. selecting crops is soil suitability. These challenges include different soil requirements based on Nitrogen, Phosphorus, and Potassium content in the soil, different PH values of soil, and variability in the climate of a region based on humidity, rainfall, and temperature of a particular region. Therefore, selecting the best crop for a farm is a tedious job for farmers nowadays and it requires careful consideration of several factors.
This model can be used by farmers for predicting the best crop for their farm in order to maximize their farm yield and also their profits through farming. This model does not ask for personalized data such as name, age, gender, religion, address, etc. You can use this web application anytime and the model will give the best crop based on the soil content and climate-oriented factors.